finance_bot/requirements.txt
anyio==4.6.0 ; python_version >= "3.12" and python_version < "4.0"
beancount==2.3.6 ; python_version >= "3.12" and python_version < "4.0"
beautifulsoup4==4.12.3 ; python_version >= "3.12" and python_version < "4.0"
bottle==0.13.1 ; python_version >= "3.12" and python_version < "4.0"
cachetools==5.5.0 ; python_version >= "3.12" and python_version < "4.0"
certifi==2024.8.30 ; python_version >= "3.12" and python_version < "4.0"
chardet==5.2.0 ; python_version >= "3.12" and python_version < "4.0"
charset-normalizer==3.3.2 ; python_version >= "3.12" and python_version < "4.0"
colorama==0.4.6 ; python_version >= "3.12" and python_version < "4.0" and sys_platform == "win32"
google-api-core==2.20.0 ; python_version >= "3.12" and python_version < "4.0"
google-api-python-client==2.146.0 ; python_version >= "3.12" and python_version < "4.0"
google-auth-httplib2==0.2.0 ; python_version >= "3.12" and python_version < "4.0"
google-auth==2.35.0 ; python_version >= "3.12" and python_version < "4.0"
googleapis-common-protos==1.65.0 ; python_version >= "3.12" and python_version < "4.0"
h11==0.14.0 ; python_version >= "3.12" and python_version < "4.0"
httpcore==0.17.3 ; python_version >= "3.12" and python_version < "4.0"
httplib2==0.22.0 ; python_version >= "3.12" and python_version < "4.0"
httpx==0.24.1 ; python_version >= "3.12" and python_version < "4.0"
idna==3.10 ; python_version >= "3.12" and python_version < "4.0"
iniconfig==2.0.0 ; python_version >= "3.12" and python_version < "4.0"
lxml==5.3.0 ; python_version >= "3.12" and python_version < "4.0"
packaging==24.1 ; python_version >= "3.12" and python_version < "4.0"
pdfminer2==20151206 ; python_version >= "3.12" and python_version < "4.0"
pluggy==1.5.0 ; python_version >= "3.12" and python_version < "4.0"
ply==3.11 ; python_version >= "3.12" and python_version < "4.0"
proto-plus==1.24.0 ; python_version >= "3.12" and python_version < "4.0"
protobuf==5.28.2 ; python_version >= "3.12" and python_version < "4.0"
pyasn1-modules==0.4.1 ; python_version >= "3.12" and python_version < "4.0"
pyasn1==0.6.1 ; python_version >= "3.12" and python_version < "4.0"
pyparsing==3.1.4 ; python_version >= "3.12" and python_version < "4.0"
pytest==8.3.3 ; python_version >= "3.12" and python_version < "4.0"
python-dateutil==2.9.0.post0 ; python_version >= "3.12" and python_version < "4.0"
python-magic==0.4.27 ; python_version >= "3.12" and python_version < "4.0" and sys_platform != "win32"
python-telegram-bot==20.4 ; python_version >= "3.12" and python_version < "4.0"
requests==2.32.3 ; python_version >= "3.12" and python_version < "4.0"
rsa==4.9 ; python_version >= "3.12" and python_version < "4"
six==1.16.0 ; python_version >= "3.12" and python_version < "4.0"
sniffio==1.3.1 ; python_version >= "3.12" and python_version < "4.0"
soupsieve==2.6 ; python_version >= "3.12" and python_version < "4.0"
uritemplate==4.1.1 ; python_version >= "3.12" and python_version < "4.0"
urllib3==2.2.3 ; python_version >= "3.12" and python_version < "4.0"
